requirements.txt
streamlit==1.33.0
streamlit-option-menu==0.3.12
firebase-admin==6.2.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
spacy==3.7.4
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.1
plotly==5.20.0
altair==5.3.0
beautifulsoup4==4.12.3
nltk==3.8.1
Pillow==10.3.0
